+++ src/Commands/CreateEditViewCommand.php
@@ -48,9 +48,9 @@ class CreateEditViewCommand extends ViewsCommandBase
     {
         $input = $this->getCommandInput();
         $resources = Resource::fromFile($input->resourceFile, $input->languageFileName);
-        $destenationFile = $this->getDestinationViewFullname($input->viewsDirectory, $input->prefix, 'edit');
+        $destenationFile = $this->getDestinationViewFullname($input->viewsDirectory, $input->prefix);
 
-        if ($this->canCreateView($destenationFile, $input->force, $resources->fields)) {
+        if ($this->canCreateView($destenationFile, $input->force, $resources)) {
             $stub = $this->getStub();
 
             $this->createLanguageFile($input->languageFileName, $input->resourceFile, $input->modelName)